requirements.txt
PyYAML>=6.0
openai>=1.0.0
more-itertools>=10.0.0
cryptography>=3.0.0
pathlib2>=2.3.0; python_version < '3.4'